cmd/webhook/provider.go: Fix wrong handling of TXT record values
Earlier there was a problem with extra double quotes seemingly being added by
external-dns and Tidy not accepting it. The workaround being to remove any
outer double quotes on TXT records and re-adding them when serving them back to
external-dns, though this breaks external-dns.

The solution it seems is that the double quotes are removed when creating the
record and nothing else. External-dns seems no to complain when these values
are read back even though the value has technically changed a bit.

Maybe I'm overlooking something but this seems to work.