cmd/webhook/provider.go: Fix annotation key
Seems the annotation webhook-tidy-description is being translated to the
providerSpecification webhook/tidy-description by external-dns. Hence the key
to match it against here is changed. It does fix the problem for the single
case we have in Netic but something seems broken in other places where it
doesn't work. The issue should probably be followed up on.