Use json.Number in place of int where appropiate in Tidy DNS interface
With the Tidy DNS server being written in Perl which has a weak type system,
we have sometimes seen json fields being returned as numbers and othertimes
stringyfied numbers. It's been discussed that supporting numbered values being
either type is probably a good idea as insuring against this is troublesome in
Perl.