cmd/webhook/provider.go: Fix TXT record leak
For some time it's been a problem with TXT records associated with A and
AAAA records, not being cleaned up on deletion. This is fixed by
cleaning the target string for wrapping "" which External DNS likes to
use but is not supported by TidyDNS.

A couple of spelling mistakes is also corrected.